+++ examples/sign_submit_transaction_sandnet.rs
@@ -1,5 +1,5 @@
 use algonaut::core::{Address, MicroAlgos};
-use algonaut::transaction::{Payment, Txn};
+use algonaut::transaction::{Pay, Txn};
 use algonaut::{Algod, Kmd};
 use dotenv::dotenv;
 use std::env;
@@ -9,13 +9,14 @@ fn main() -> Result<(), Box<dyn Error>> {
     // load variables in .env
     dotenv().ok();
 
+    // kmd manages wallets and accounts
     let kmd = Kmd::new()
         .bind(env::var("KMD_URL")?.as_ref())
         .auth(env::var("KMD_TOKEN")?.as_ref())
         .client_v1()?;
 
+    // first we obtain a handle to our wallet
     let list_response = kmd.list_wallets()?;
-
     let wallet_id = match list_response
         .wallets
         .into_iter()
@@ -24,48 +25,46 @@ fn main() -> Result<(), Box<dyn Error>> {
         Some(wallet) => wallet.id,
         None => return Err("Wallet not found".into()),
     };
-    println!("Wallet: {}", wallet_id);
-
     let init_response = kmd.init_wallet_handle(&wallet_id, "")?;
     let wallet_handle_token = init_response.wallet_handle_token;
+    println!("Wallet Handle: {}", wallet_handle_token);
 
+    // an account with some funds in our sandbox
     let from_address = Address::from_string(env::var("ACCOUNT")?.as_ref())?;
-    println!("Sender: {:#?}", from_address);
+    println!("Sender: {:?}", from_address);
 
     let to_address =
         Address::from_string("2FMLYJHYQWRHMFKRHKTKX5UNB5DGO65U57O3YVLWUJWKRE4YYJYC2CWWBY")?;
-    println!("Receiver: {:#?}", to_address);
+    println!("Receiver: {:?}", to_address);
 
+    // algod has a convenient method that retrieves basic information for a transaction
     let algod = Algod::new()
         .bind(env::var("ALGOD_URL")?.as_ref())
         .auth(env::var("ALGOD_TOKEN")?.as_ref())
         .client_v1()?;
 
     let params = algod.transaction_params()?;
-    let genesis_id = params.genesis_id;
-    let genesis_hash = params.genesis_hash;
 
-    let payment = Payment {
-        amount: MicroAlgos(123_000),
-        receiver: to_address,
-        close_remainder_to: None,
-    };
-    println!("Payment: {:#?}", payment);
-
+    // we are ready to build the transaction
     let t = Txn::new()
         .sender(from_address)
         .first_valid(params.last_round)
         .last_valid(params.last_round + 1000)
-        .genesis_id(genesis_id)
-        .genesis_hash(genesis_hash)
+        .genesis_id(params.genesis_id)
+        .genesis_hash(params.genesis_hash)
         .fee(MicroAlgos(10_000))
-        .payment(payment)
+        .payment(
+            Pay::new()
+                .amount(MicroAlgos(123_456))
+                .to(to_address)
+                .build(),
+        )
         .build();
 
+    // we need to sign the transaction to prove that we own the sender address
     let sign_response = kmd.sign_transaction(&wallet_handle_token, "", &t)?;
-    println!("Signed: {:#?}", sign_response);
 
-    // Broadcast the transaction to the network
+    // broadcast the transaction to the network
     let send_response = algod.raw_transaction(&sign_response.signed_transaction)?;
 
     println!("Transaction ID: {}", send_response.tx_id);